requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
fuzzywuzzy==0.18.0
h11==0.13.0
idna==3.3
numpy==1.21.5
opencv-python==4.5.5.62
outcome==1.1.0
pandas==1.1.5
Pillow==9.0.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
termcolor==1.1.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0

loguru~=0.6.0